Fix section for performance
