+++ customtypes/home/index.json
@@ -11,6 +11,9 @@
         "fieldset": "Slice Zone",
         "config": {
           "choices": {
+            "about_with_images_grid": {
+              "type": "SharedSlice"
+            },
             "faq_with_cta": {
               "type": "SharedSlice"
             },
