+++ slices/AboutWithImagesGrid/index.tsx
@@ -1,6 +1,7 @@
 import { FC } from "react";
 import { Content } from "@prismicio/client";
-import { SliceComponentProps } from "@prismicio/react";
+import { PrismicRichText, SliceComponentProps } from "@prismicio/react";
+import Section from "@/components/Section";
 
 /**
  * Props for `AboutWithImagesGrid`.
@@ -13,20 +14,32 @@ export type AboutWithImagesGridProps =
  */
 const AboutWithImagesGrid: FC<AboutWithImagesGridProps> = ({ slice }) => {
   return (
-    <section
-      data-slice-type={slice.slice_type}
-      data-slice-variation={slice.variation}
-    >
-      Placeholder component for about_with_images_grid (variation:{" "}
-      {slice.variation}) slices.
-      <br />
-      <strong>You can edit this slice directly in your code editor.</strong>
-      {/**
-       * 💡 Use the Prismic MCP server with your code editor
-       * 📚 Docs: https://prismic.io/docs/ai#code-with-prismics-mcp-server
-       */}
-    </section>
-  );
+		<Section
+			hasBlub={false}
+			data-slice-type={slice.slice_type}
+			data-slice-variation={slice.variation}
+		>
+			<div className="flex items-center justify-center flex-col">
+				<span className="text-lg min-w-4 flex items-center justify-center overflow-hidden px-6 py-1 border rounded-full">
+					<PrismicRichText field={slice.primary.heading} />
+				</span>
+				<div className="text-center max-w-2xl text-lg shrink-0 mt-5">
+					<PrismicRichText field={slice.primary.description} />
+				</div>
+
+				<div className="mt-8 flex gap-3 relative">
+					{Array.isArray(slice.primary.cards) &&
+						slice.primary.cards.length &&
+						slice.primary.cards.map((item, index) => (
+							<div key={index} className="overflow-hidden w-full rounded-lg">
+                <img src={item.media.url} height={500} width={250} className="object-cover h-[300px]" />
+                
+							</div>
+						))}
+				</div>
+			</div>
+		</Section>
+	)
 };
 
 export default AboutWithImagesGrid;
